handle refnames made by git-1.8.3 and later
Git-1.8.3 introduced a change to the way "$Format" strings (specifically
the "%d" variable) are expanded in git-archive -generated tarballs. The
list of refs has always included both branch names and tags. But in
older versions, tags were not emitted any differently than branch names:
the list might say:

 (master, HEAD, v1.0)

Starting in 1.8.3, tags are marked with a distinctive "tag: " prefix, so
you get:

 (master, HEAD, tag: v1.0)

The versioneer code that deduces version tags from generated tarballs
was not spotting the newly-marked tags, and falling back to reporting a
full SHA1 hash as the revision, instead of the intended tag name.

This change fixes the parser and restores compatibility with newer
versions of git. Closes: #11.