PEP440-compatibility: rewrite parser, change version format
The short version should now always be PEP440-compliant, in one of the
following forms:

* TAG
* TAG+DISTANCE.gHASH[.dirty]
* 0+unknown
* 0+untagged.gHASH[.dirty]
* 0+unparseable[.dirty]

This uses PEP440's "local version" section to represent the non-tag
portions, retaining all the information from before but not triggering
"LegacyVersion" warnings. ".dirty" is used instead of "-dirty" because
"-dirty" would trigger a warning and be normalized into ".dirty"
anyways.

Thanks to @muggenhor, @sebastianneubauer, and @glennmatthews for the
many attempts to solve this which have been incorporated into this
patch.