Run pip using `python -m pip` and `rundemo` optionally using python.
For `pip` the problem in Windows was that `pip install` didn't have the correct permissions to the pip uninstall file. Using `python -m pip install` alleviates this access issue.

For `rundemo`: for some of the demo projects, this is simply a python file. In Linux you could make it runnable on its own through a shebang line (`#!/usr/bin/env python`) but this doesn't work on windows CMD or Powershell. In these cases we run `python rundemo` explicitly instead, controllable through the `use_python` argument.