MNT: Relicense to Unlicense
Versioneer was initially licensed simply "public domain". For clarity,
we chose CC0 as an actual license that puts the work into the public
domain. However, Fedora has rejected CC0 as a permissible code license
because it explicitly reserves the right to patent. Therefore we are
moving to the Unlicnese, which is simpler and has no such provision."